+++ src/entities/customer/index.ts
@@ -4,6 +4,7 @@
  *  Changes may be overwritten as part of auto-generation.
  */
 
+export * from './auth-token.js';
 export * from './customer-collection.js';
 export * from './customer.js';
 export * from './credit-balance.js';