+++ src/enums/notification-settings/index.ts
@@ -5,3 +5,4 @@
  */
 
 export * from './notification-settings-type.js';
+export * from './traffic-source.js';