+++ src/resources/notification-settings/operations/list-notification-settings-query-parameters.ts
@@ -4,9 +4,12 @@
  *  Changes may be overwritten as part of auto-generation.
  */
 
+import { type TrafficSource } from '../../../enums/index.js';
+
 export interface ListNotificationSettingsQueryParameters {
   after?: string;
   perPage?: number;
   orderBy?: string;
   active?: boolean;
+  trafficSource?: TrafficSource;
 }