+++ src/types/customer/index.ts
@@ -4,6 +4,7 @@
  *  Changes may be overwritten as part of auto-generation.
  */
 
+export * from './auth-token-response.js';
 export * from './customer-response.js';
 export * from './credit-balance-response.js';
 export * from './customer-balance.js';